Add support for fold_quantize=True (#1235)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/1235

This is needed since quant flow is moving to enable this by default. For now we have to support both.

Not turning the knob on in this diff, so we can run the CI with this flag disabled. And the next diff can enable the knob so we can test with it.

Reviewed By: mcr229

Differential Revision: D51419440

fbshipit-source-id: 7eec01b83ef81c2e72cc65386d68268204068018