Add MoE structure in the llama example (#1533)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/1533

Support MoE structure, where there can be multiple experts in the FFN layer.
The change in model.py is based on https://github.com/pytorch-labs/gpt-fast/pull/71.
Note that it's a functional verification, with random weights.
It can be successful exported and lowered to ExecuTorch.

TODO: test the runtime side.

Reviewed By: larryliu0820

Differential Revision: D52543030

fbshipit-source-id: 5d4220f1e8ea9eb1e4be398fe2a47bfb0b89c975