Fix emission of weights that are views
Summary: If a weight was a view of some larger data blob we would get a disconnect between the logical nbytes and the actual serialized storage. Thats not great and was upsetting deserialization. The .clone() should prevent it. If anyone knows how to do to this without the overhead of clone (as in just getting a pointer to relevant block of storage) Id be happy to do that instead.

Reviewed By: zhxchen17, mcremon-meta

Differential Revision: D53010225

fbshipit-source-id: 70ff705a01c84a7b300e1743af2939ae15dc6939