Add mixed mm op (#1791)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/1791

This op performs matrix multiplication between a Float/Half `input` tensor, and a quantized int8 `weight` tensor. The op takes a `weight_scales` tensor which must be of the same dtype than the `input`, and an optional `weight_zero_points` tensor, which also must have same dtype than the input.
The output tensor must also be of the same dtype than the input.

Reviewed By: mikekgfb

Differential Revision: D53283501

fbshipit-source-id: 0d85c666f3d95ec49cb2c18c2d57c5312cd843b0