Serialize the flag of appending EOS tokens into the model (#1832)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/1832

Previously we need to pass in a flag into runner so that it conditionally append EOS tokens to prompt tokens.

This is flexible for the runtime to use llama2 7b, where the user can let llama2 finish the sentence if no EOS is in the prompt, or have a conversation where EOS is in the prompt.

However it is harder to control and is a footgun for developers to debug. This diff serialize the flag into the model and simplify the runner logic.

Reviewed By: shoumikhin, mikekgfb

Differential Revision: D53361647

fbshipit-source-id: 82d3b8a3540f6ed7bd7f00ff99a3702f9a92cbdc