MmapDataLoader can handle non-page-aligned offsets (#1975)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/1975

Instead of requiring that load offsets are aligned to the size of the OS page, map in the set of pages that overlap the requested offset + size.

This will let us load .pte files with 4096-aligned segments onto iOS devices, which use 16384-byte OS pages.

It also means that we could reduce the segment alignment to `max(delegate_alignment, constant_tensor_alignment)` to reduce the size of .pte files.

Reviewed By: shoumikhin

Differential Revision: D53772818

fbshipit-source-id: d08cef5345d1abaad060a6b40cd0db18a7f1b94b