Remove unused variables in executorch/examples/sdk/sdk_example_runner/sdk_example_runner.cpp (#2007)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/2007

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D53779562

fbshipit-source-id: ba6fd427fe8d4dcba9020d858122a687e9e9bf0e