duplicate schema for AoT and Runtime (Temporary) (#1877)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/1877

Preliminary setup for updating schema and XNNPACK Backend Runtime for ASR Backwards Compatibility.

The key to this is doing a manual fix by following these steps:
1. Make Runtime only changes to support XN00 and XN01
1a. This runtime should be able to support models generated using the old schema (XN00) and models using the new schema XN01. This should only change how the model is deserialized and executed
2. Update model generation to generate new XN01 models
2a. This should include AoT only changes, and should only change how the model artifact is serialized.
3. Remove the support for XN00 in the runtime so that only XN01 is now supported.

The above steps must be performed 1-by-1 in which after landing step n, we wait for some specified amount of time and end-to-end testing is finished, before attempting to land the next step.

The purpose of this diff is to facilitate each change being contained within their own step. Because we wish to make runtime only changes and AoT-only changes for each distinct step, we split the schema.

This way at step 1, we modify only the runtime schema to support XN01 and XN00 models and ensure that all schema related changes are only applied runtime.

At step 2, we then modify the AoT Schema to our desired final schema, this schema will serialize only XN01 models. We ensures that the schema changes done at this step only affect the AoT model generation.

Finally at step three, when we are generating and delivering the XN01 model, we delete the runtime schema (support for XN00 and XN01) and move fully onto the AoT schema which only supports XN01.

Reviewed By: digantdesai

Differential Revision: D53454517

fbshipit-source-id: a1ddc4a696b7970323f6ad02e7e72e85520be387