Fix of the test-arm-backend-delegation flow (#2143)
Summary:
There are a few oddities in the test setup, hopefully this fixes https://github.com/pytorch/executorch/issues/2011 and fixes https://github.com/pytorch/executorch/issues/1947.

The change is to use a per-test-run build of the setup.sh / SDK components rather than one in /opt/arm-sdk - the setup.sh inlcudes a pip install of a vela executable into the venv path which was not correctly carried through into the test run.

Once this works we can remove the Docker stage which is running setup.sh

Pull Request resolved: https://github.com/pytorch/executorch/pull/2143

Reviewed By: digantdesai

Differential Revision: D54310891

Pulled By: mergennachin

fbshipit-source-id: eef28cd13e73bb44d84910b49896e43a2a407550