Update kernel utils for non-fatal (#1770)
Summary:
Note to reviewers: Because the changeset for non-fatal touches so many files, I've split into a number of diffs. The tests will not all pass until the final diff, given that non-fatal tests are switched on in the first diff but tests are not updated until the respective later diff. See revision 6.6 of D53251417 to observe the full change-set passing CI.

Pull Request resolved: https://github.com/pytorch/executorch/pull/1770

Test Plan: buck2 test mode/dev-nosan kernels/test:

Reviewed By: manuelcandales

Differential Revision: D53251417

Pulled By: GregoryComer

fbshipit-source-id: 6cb330af0acffa13acc45a2dbbe99027627071c9