Update kernels for non-fatal (part 4) (#2107)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/2107

Update kernels and tests starting with 'd'.

Note to reviewers: Because the changeset for non-fatal touches so many files, I've split into a number of diffs. The tests will not all pass until the final diff, given that non-fatal tests are switched on in the first diff but tests are not updated until the respective later diff. See revision 6.6 of D53251417 to observe the full change-set passing CI.

bypass-github-export-checks
bypass-github-executorch-ci-checks

Reviewed By: kirklandsign

Differential Revision: D54181994

fbshipit-source-id: d5002f2b61119dab379911fcb88b11e0a3c04f0e