Update kernels for non-fatal (part 17) (#2114)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/2114

Update kernels and tests starting with 't'.

Note to reviewers: Because the changeset for non-fatal touches so many files, I've split into a number of diffs. The tests will not all pass until the final diff, given that non-fatal tests are switched on in the first diff but tests are not updated until the respective later diff. See revision 6.6 of D53251417 to observe the full change-set passing CI.

bypass-github-export-checks
bypass-github-executorch-ci-checks

Reviewed By: manuelcandales

Differential Revision: D54182490

fbshipit-source-id: 5718f5ab35e5bd682430b6559896de4d6d83a2e0