Optimize memory when exporting. (#2341)
Summary:
CoreML delegate when converting the `mlpackage` to an in-memory representation reads the directory contents in the `mlpackage` directory and writes it to a memory buffer this is expensive and the copying is not required.

The change in the PR addresses by memory mapping the file contents to fixed addresses in the allocated memory, no memory is dirtied and the OS will bring the pages only when the bytes are read.

Pull Request resolved: https://github.com/pytorch/executorch/pull/2341

Reviewed By: digantdesai

Differential Revision: D54997652

Pulled By: shoumikhin

fbshipit-source-id: 2f479be5241f6d2c60005a2f6410514ee967a46e