Don't partition non_packable constants (#2423)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/2423

When constant data is not packed(constant data is packed for weights and biases like those supplied to convolution, bmm, linear, etc.), it is not copied by XNNPACK runtime. As a result, when the flatbuffer payload is freed after init, so are the pointers given to XNNPACK.

We can solve this by unpartitioning constant data nodes which aren't packed. As a result they are owned by ExecuTorch and are not freed by xnnpack delegate after init

bypass-github-pytorch-ci-checks
bypass-github-export-checks
bypass-github-executorch-ci-checks

Reviewed By: digantdesai, cccclai

Differential Revision: D54877319

fbshipit-source-id: d54c477abc1a4561bbd19db1ffef65f33e7ea742