Restore CMake configure comment (#2723)
Summary:
A comment at the top of the top-level CMakeLists.txt file was removed previously as part of an update to automatically download buck. I removed more than was needed. This change restores the comment (minus the explicit buck path).

Pull Request resolved: https://github.com/pytorch/executorch/pull/2723

Test Plan: This is a comment-only change. I did configure and build on linux x86-64 to validate syntax.

Reviewed By: mcr229

Differential Revision: D55445749

Pulled By: GregoryComer

fbshipit-source-id: ff8cb601e71be4231f4e853b993bad9f54ea7144