Error out when token is outside of vocab size (#3535)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/3535

Ideally it shouldn't happen, but if we post process the weight somehow too much it might happen. In Android, it just seg fault directly if it's outside of the range without error message. After this change, it's clearer:
```
E 00:00:00.180911 executorch:bpe_tokenizer.cpp:155] token 18446744073709551615 is out side of vacab range 512
Aborted
```

Reviewed By: larryliu0820

Differential Revision: D57057026

fbshipit-source-id: 838260d60b75e7c392d7f496d7cdf6f81957f56c