Define custom op `conv_with_clamp` through python (#3886)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/3886

We implement a custom op `torch.ops.et_vk.conv_with_clamp` through `torch.convolution` and `torch.clamp` and test it in a python script `test_custom_ops.py`. This is to prepare the fusion of `conv` and `relu, hardtanh` on Vulkan. See the following diffs for the implementation.

Reviewed By: SS-JIA

Differential Revision: D58173778

fbshipit-source-id: 2f5ea16cf217f559bdb9b345cbcb5d45810cab0c