use _to_copy or _to_dim_order_copy in I64toI32 pass (#3968)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/3968

When converting an ExportedProgram to EdgeProgramManager via to_edge, _to_dim_order_copy or _to_copy will be used depending on the passed EdgeCompileConfig having _skip_dim_order set to true or false. For example:
```
compile_config=EdgeCompileConfig(
    _check_ir_validity=False,
    _skip_dim_order=True,
),
```
The newer _to_dim_order_copy op was added in D53747744/https://github.com/pytorch/executorch/pull/1970.

Adding a param to the pass init to determine if we use the newer op or not.

Reviewed By: copyrightly

Differential Revision: D58395616

fbshipit-source-id: 8f5289f8ed62a5c3814ec32b91fb055a10944fdb