Get device information from OpenCL (#4210)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/4210

Some of the device's properties, like the number of available Streaming Processors, or the total size of the cache, are not supported by Vulkan as direct queries. These variables are essential for ArchProbe's algorithms, so we need to do some preprocessing with OpenCL to extract these values, before moving to the Vulkan implementation.

This diff moves the GPUInfo implementation into a class to have better control of internal variables like the device's properties, and obtains important limits that will be used in the following diffs, like the device's SM Count, Cache size and number of logic threads.

Reviewed By: jorgep31415

Differential Revision: D59636879

fbshipit-source-id: 1b9aa3a4ce48f360526e09576864514b2ac35429