Update serializer to sync with export serializer (#4264)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/4264

This diff brings in the latest export serializer changes from `torch/_export/serde` and then fixes the exir serializer to use these changes. This is a temporary workaround until the serialization extensiblity that zhxchen17 is working on is completed. Then we'll not have to copy over changes from export serializer manually, instead we'll just extend it to handle things like delegate calls and other edge dialect specific things.

For now we need this to unblock ASR use case and hence i'm manually syncing and updating for now.

Reviewed By: JacobSzwejbka

Differential Revision: D57071033

fbshipit-source-id: 4408e2a0740b661e3a3555f800d1567ef10d4ea8