Gh/larryliu0820/46/base (#4643)
* [llava][13/N] Move metadata util to a separate header for reuse

As titled.

[ghstack-poisoned]

* Update on "[llava][13/N] Move metadata util to a separate header for reuse"


As titled.

[ghstack-poisoned]

* Update on "[llava][13/N] Move metadata util to a separate header for reuse"


As titled.

[ghstack-poisoned]

* Update on "[llava][13/N] Move metadata util to a separate header for reuse"


As titled.

[ghstack-poisoned]

* Update on "[llava][13/N] Move metadata util to a separate header for reuse"


As titled.

[ghstack-poisoned]

* [llava][14/N] Refactor runner prefill() and run_model_step()

This refactoring is needed in order to extract out prefill() and
run_model_step() out from runner so that these APIs become replaceable
and easy to plugin and use.

* prefill():
For the case where parallel prefill is enabled or not using kv cache,
the model is able to accept a large block (more than 1) of tokens.

For the other case where we have kv cache but parallel prefill is not
enabled, we can only feed in 1 token every time.

* run_model_step():
This function should not update the input. Instead it should run the
model differently, depending on whether kv cache is enabled. This should
return the next token directly. All the input update needs to happen in
the generation loop.

[ghstack-poisoned]

* Update on "[llava][14/N] Refactor runner prefill() and run_model_step()"


This refactoring is needed in order to extract out prefill() and
run_model_step() out from runner so that these APIs become replaceable
and easy to plugin and use.

* prefill():
For the case where parallel prefill is enabled or not using kv cache,
the model is able to accept a large block (more than 1) of tokens.

For the other case where we have kv cache but parallel prefill is not
enabled, we can only feed in 1 token every time.

* run_model_step():
This function should not update the input. Instead it should run the
model differently, depending on whether kv cache is enabled. This should
return the next token directly. All the input update needs to happen in
the generation loop.

[ghstack-poisoned]

* Update on "[llava][14/N] Refactor runner prefill() and run_model_step()"


This refactoring is needed in order to extract out prefill() and
run_model_step() out from runner so that these APIs become replaceable
and easy to plugin and use.

* prefill():
For the case where parallel prefill is enabled or not using kv cache,
the model is able to accept a large block (more than 1) of tokens.

For the other case where we have kv cache but parallel prefill is not
enabled, we can only feed in 1 token every time.

* run_model_step():
This function should not update the input. Instead it should run the
model differently, depending on whether kv cache is enabled. This should
return the next token directly. All the input update needs to happen in
the generation loop.

[ghstack-poisoned]

* [llava][15/N] Extract out text decoder runner

Last PR #4556 refactored run_model_step() so that it is suitable to be
extracted out as a separate class.

This new `TextDecoderRunner` provides 2 APIs:

* step(tokens, start_pos)
This API takes one or more tokens with start_pos and feed them into Module. Return a tensor of logits.

* logits_to_token(logits)
This API samples the result and returns a token. We don't expect this
logic to change across different runners.

[ghstack-poisoned]

* Update on "[llava][14/N] Refactor runner prefill() and run_model_step()"


This refactoring is needed in order to extract out prefill() and
run_model_step() out from runner so that these APIs become replaceable
and easy to plugin and use.

* prefill():
For the case where parallel prefill is enabled or not using kv cache,
the model is able to accept a large block (more than 1) of tokens.

For the other case where we have kv cache but parallel prefill is not
enabled, we can only feed in 1 token every time.

* run_model_step():
This function should not update the input. Instead it should run the
model differently, depending on whether kv cache is enabled. This should
return the next token directly. All the input update needs to happen in
the generation loop.

Differential Revision: [D60840327](https://our.internmc.facebook.com/intern/diff/D60840327)

[ghstack-poisoned]

* Update on "[llava][15/N] Extract out text decoder runner"


Last PR #4556 refactored run_model_step() so that it is suitable to be
extracted out as a separate class.

This new `TextDecoderRunner` provides 2 APIs:

* step(tokens, start_pos)
This API takes one or more tokens with start_pos and feed them into Module. Return a tensor of logits.

* logits_to_token(logits)
This API samples the result and returns a token. We don't expect this
logic to change across different runners.

Differential Revision: [D60856571](https://our.internmc.facebook.com/intern/diff/D60856571)

[ghstack-poisoned]

* Update base for Update on "[llava][15/N] Extract out text decoder runner"


Last PR #4556 refactored run_model_step() so that it is suitable to be
extracted out as a separate class.

This new `TextDecoderRunner` provides 2 APIs:

* step(tokens, start_pos)
This API takes one or more tokens with start_pos and feed them into Module. Return a tensor of logits.

* logits_to_token(logits)
This API samples the result and returns a token. We don't expect this
logic to change across different runners.

Differential Revision: [D60856571](https://our.internmc.facebook.com/intern/diff/D60856571)

[ghstack-poisoned]

* [llava][16/N] Extract out prefill logic into a new class

Depends on whether parallel or sequential prefill is chosen, prefill()
calls `TextDecoderRunner.step()` to prefill prompt tokens to LLM.

[ghstack-poisoned]

* Update base for Update on "[llava][16/N] Extract out prefill logic into a new class"


Depends on whether parallel or sequential prefill is chosen, prefill()
calls `TextDecoderRunner.step()` to prefill prompt tokens to LLM.

Differential Revision: [D60927756](https://our.internmc.facebook.com/intern/diff/D60927756)

[ghstack-poisoned]

* Update base for Update on "[llava][16/N] Extract out prefill logic into a new class"


Depends on whether parallel or sequential prefill is chosen, prefill()
calls `TextDecoderRunner.step()` to prefill prompt tokens to LLM.

Differential Revision: [D60927756](https://our.internmc.facebook.com/intern/diff/D60927756)

[ghstack-poisoned]

* Update base for Update on "[llava][16/N] Extract out prefill logic into a new class"


Depends on whether parallel or sequential prefill is chosen, prefill()
calls `TextDecoderRunner.step()` to prefill prompt tokens to LLM.

Differential Revision: [D60927756](https://our.internmc.facebook.com/intern/diff/D60927756)

[ghstack-poisoned]

* Update base for Update on "[llava][17/N] Move util.h into /e/llm/runner"


So that it can be reused

Differential Revision: [D60938984](https://our.internmc.facebook.com/intern/diff/D60938984)

[ghstack-poisoned]

* Update base for Update on "[llava][17/N] Move util.h into /e/llm/runner"


So that it can be reused

Differential Revision: [D60938984](https://our.internmc.facebook.com/intern/diff/D60938984)

[ghstack-poisoned]

* [llava][17/N] Move util.h into /e/llm/runner

Differential Revision: D60938984

Pull Request resolved: https://github.com/pytorch/executorch/pull/4588