[XNNPACK][Partitioner] enable src based partitioner (#4795)
In order to maintain parity with the current to_edge and to_backend lowering flow, we need to support source based partitioning.

We apply source-based partitioning to the AddMMConfig to partition all the nodes surrounding addmm so that it can be recomposed internally. While this is fine for to_edge and to_backend flow. For more robust flow, we will not have to use this when running to_edge_transform_and_lower.

Co-authored-by: Max Ren <maxren@meta.com>

Pull Request resolved: https://github.com/pytorch/executorch/pull/4762