vTensor cleanup 2/N - remove `discard_and_reallocate` API (#5422)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/5422

## Context

Remove the unused `discard_and_reallocate()` API. The purpose of this AP is to allow a `vTensor` instance to request a new memory allocation when being resized to a larger size than what the current memory allocation can support.

This was originally intended to support unbounded dynamic shapes in the ExecuTorch Vulkan delegate, but it is unlikely that this functionality will ever be used. The reason is that re-allocating memory would require re-encoding the command buffer, which goes against the execution philosophy of the Vulkan delegate (encode command buffer once, submit repeatedly).
ghstack-source-id: 243061031
exported-using-ghexport

Reviewed By: jorgep31415

Differential Revision: D62878649

fbshipit-source-id: e293fe2dfcbbec8c615370a370db14424b971369