Add New Ethos-U85 support to Arm aot_arm_compiler (#5345)
Summary:
Some changes to enable using the new compiler flow and backend for both Ethos-U55 and Ethos-U85.

This change means the run.sh flow uses the new compiler for Ethos-U55.

It enables Ethos-U85 to be targeted for compilation, but this feature isn't used in run.sh yet.

Pull Request resolved: https://github.com/pytorch/executorch/pull/5345

Reviewed By: manuelcandales

Differential Revision: D62874758

Pulled By: digantdesai

fbshipit-source-id: fd9921b7dd656a0087aaffde5986f4bdc320742c