Implement slice as a view (#5590)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/5590

## Context

TSIA. Implement slice as a view operator. This is only valid under the following conditions:

* All dims preceding the sliced dim in the dim order have a size of 1
* start is 0
* step is 1

The reasoning for these restrictions is so that the offset of the slice view with respect to the source buffer is 0. More details are in the comments.

To test the operator effectively, this diff also extends the test codegen to handle multiple test suites for one operator, each with a different configuration.
ghstack-source-id: 244431147
exported-using-ghexport

Reviewed By: jorgep31415

Differential Revision: D61666462

fbshipit-source-id: e4645ec672be0699c88eb1bb88fdef5b4e5cfdb1