Upload Android benchmark results to OSS benchmark database (#5808)
Summary:
This PR adds a job to upload Android benchmark results to the benchmark database.  It transforms the `benchmark_results.json` file slightly to fit into the current schema.  We are going to have a better schema soon https://fburl.com/gdoc/ossgtvte, but landing this first would unblock the work on building the dashboard before the launch.  Updating the schema can be done later.

* The job processes what it finds, so if one model fails, the rest will still be uploaded.
* I will follow up with another PR for iOS later. No need to wait for the TPS metric there, we'll upload what available first.

There are still some TODO pending:

* https://github.com/pytorch/test-infra/pull/5742
* https://github.com/pytorch-labs/pytorch-gha-infra/pull/483

But the structure of the CI job is ready to review.

Pull Request resolved: https://github.com/pytorch/executorch/pull/5808

Reviewed By: guangy10, kirklandsign

Differential Revision: D63869876

Pulled By: huydhn

fbshipit-source-id: f9bf85c9599fafbfcc300d47e4307230c46b16db