[ET-VK] Replace Uniform buffers with push constants for permute op (#7347)
* [ET-VK] Replace Uniform buffers with push constants for  binary op

Pull Request resolved: https://github.com/pytorch/executorch/pull/7230

This diff replaces uniform buffers with push constants for binary op in the Vulkan backend of Executorch. The changes include updating the GLSL code to use push constants instead of uniform buffers and updating the C++ code to pass the sizes as push constants to the shader.
ghstack-source-id: 258575398
@exported-using-ghexport

Differential Revision: [D66853542](https://our.internmc.facebook.com/intern/diff/D66853542/)

* [ET-VK] Replace Uniform buffers with push constants for permute op

Pull Request resolved: https://github.com/pytorch/executorch/pull/7231

This diff replaces uniform buffers with push constants for permute op in the Vulkan backend of Executorch. The changes include updating the GLSL code to use push constants instead of uniform buffers and updating the C++ code to pass the sizes as push constants to the shader.
ghstack-source-id: 258575396
@exported-using-ghexport

Differential Revision: [D66890825](https://our.internmc.facebook.com/intern/diff/D66890825/)

---------

Co-authored-by: Vivek Trivedi <5340687+trivedivivek@users.noreply.github.com>