Arm backend: Add check to not partition float inputs for BI (#8681)
Add check to not partition float inputs for BI

Floats are not supported in TOSA BI profile. Some supported operators
are only quantized if the previous node was quantized. In practice,
this means that if an unsupported operator preceeds such an operator,
it will not be quantized and the input will be a float. This will
likely lead to an assertion error or invalid TOSA graph. This patch
aims to detect such nodes, and to reject them.

Signed-off-by: Oscar Andersson <oscar.andersson@arm.com>