[ET-VK] Adding a workgroup class to VecUtils (#8669)
Pull Request resolved: https://github.com/pytorch/executorch/pull/8632

This diff adds a new class called `WorkgroupSize` to the `VecUtils` header file. The `WorkgroupSize` class takes three `uint32_t` values as parameters and stores them in a single `uint32_t` variable using bitwise operations. This class is used in the Vulkan backend to specify the size of a workgroup for a given operation.
ghstack-source-id: 268172661
@exported-using-ghexport

Differential Revision: [D70021019](https://our.internmc.facebook.com/intern/diff/D70021019/)

Co-authored-by: Vivek Trivedi <5340687+trivedivivek@users.noreply.github.com>