Arm backend: Make passes preserve and update node metadata (#9362)
When creating or updating nodes in passes, the metadata is not preserved
nor updated correctly. This patch adds an ArmPass base class which may
update the node metadata if super().call_operator(update=True) is used.
It also adds functionality to arm_pass_utils.create_node() to update the
node metadata. It will only update the 'stack_trace' field. All the
other fields will be preserved from the original node.


Signed-off-by: Oscar Andersson <oscar.andersson@arm.com>