Arm backend: Update partitioner de-tagging iteration order (#10813)
When de-tagging tagged/partitioned nodes, the order in which we iterate
the nodes will matter as each node will make a decision based on if the
parent node is partitoned or has a compatible dtype. This change makes
sure we iterate the nodes in the order of the original graph_module.

Signed-off-by: Oscar Andersson <oscar.andersson@arm.com>