Arm backend: Make the CI green by not testing Dedicated_Sram for the Ethos-U85 (#10973)
Temporary solution to the problem in
https://github.com/pytorch/executorch/pull/10958 The
arm_executor_runner.cpp need to declare the ethosu_fast_scratch array
and pass it onto to the EthosUBackend.cpp. It is important that for
Shared_Sram, the ethosu_fast_scratch is nullptr and for Dedicated_Sram
it points to the fast memory array.