Arm backend: Make all partitioned Batchnorm2D to conv (#11148)
Batchnorm2Ds can be fused with preceeding convs.
For float, this is done in the FuseBatchnorm2dPass which with this patch
is updated to insert 'transparent' convs before stand-alone bns. This
way all bns are fused with convs instead of decomposed, so
DecomposeBatchNorm2DPass is removed.

For quantized, this is done by _fuse_conv_bn_(model) in the prepare_p2e
function, before the transform_for_annotation pipeline. This does not
allow us to modify the graph before the transform. Running our own
DecomposeBatchNorm2dPass before quantization on the other hand requires
modifying constant tensors before the export step which is a non trival
effort.

Since stand-alone batchnorms are not likely to be seen in real graphs,
instead simply do not accept quantized stand-alone batchnorms for now.
tosa_supported_operators is updated to give a better explanation in this
case.

Additionally refactors batchnorm unittest no be more brief and to more
properly reflect what works and not.

Signed-off-by: Adrian Lundell <adrian.lundell@arm.com>