Arm backend: Remove special host build of flatcc (#11278)
### Summary
This was a workaround as host version of flattcc was not generated from
the build as it should and you got the target version. Later when
flattcc was needed on the host it was the wrong version and you got an
error.
Now this has been solved in the project and flattcc is only build for
the host so there is no need to force a special build like this anymore.

### Test plan
Tested in local build and on GitHub

Signed-off-by: Zingo Andersen <zingo.andersen@arm.com>