Prim ops replace abort with error code (#11985)
### Summary
We want to minimize the scenarios where ET fails fatally. If we need to
check a precondition inside a kernel rather then calling ET_CHECK which
internally dispatches to ABORT we should call ET_KERNEL_CHECK which sets
an error state and returns.

### Test plan
Replaced TestNegScalarWithTensorDies with TestNegScalarWithTensorFails.
Before, it would expect an abort when negating a tensor. The new test
expects a failure state instead.

The test can be run with `test/run_oss_cpp_tests.sh`