Arm backend: Split executor runner into init + run (#12162)
Refactor the main function of arm_executor_runner.cpp by extracting code
that is related to initializing and running a model into two separate
helper functions (runner_init and RunnerContext::run). A new struct
called RunnerContext is introduced to store data/context required to run
the model.

### Test plan
The code is tested in the CI. It's a refactoring patch, so no new
functionality added/removed.

Co-authored-by: Martin Lindström <Martin.Lindstroem@arm.com>