Resubmission of Arm backend: Split executor runner into init + run (#12197)
The PR #12162 got reverted with #12171. This PR is a resubmission of the
patch.

Refactor the main function of arm_executor_runner.cpp by extracting code
that is related to initializing and running a model into two separate
helper functions (runner_init and RunnerContext::run). A new struct
called RunnerContext is introduced to store data/context required to run
the model.

### Test plan
Arm's CI runs and tests the changed code.

Co-authored-by: Martin Lindström <Martin.Lindstroem@arm.com>