Define edge exception and preserve as global (#12300)
Summary: There is a list of ops provided as arguments to the
to_edge_with_preserved_ops and EdgeCompileConfig. This PR pulls these
lists out of the function and hoists it above the function as a global
variable, allowing it to be imported into other code to align compiler
behavior between different callsites.

Differential Revision: D78008871