PAL File for Arm Zephyr Executor Runner (#12348)
### Summary
Platform file for building Zephyr applications that use Executorch as a
module. This is a simplified version of the file used in a Zephyr
development branch
[here](https://github.com/BujSet/zephyr/blob/executorch-module-integration/samples/modules/executorch/arm/hello_world/src/arm_zephyr_pal.hpp).

In the future, it may be prudent to have a PAL file for arm bare metal
(I imagine this would require a simple splicing of he PAL functions from
the `arm_executor_runner.cpp` file). Leaving that task for a separate
PR.

### Test plan
Tested against the `arm_executor_runner.cpp` that is used in the
[Arm-EthosU
tutorial](https://docs.pytorch.org/executorch/main/tutorial-arm-ethos-u.html).
Was able to successfully build the Arm Executor Runner, and run the
simple add model.