[ET-VK] 3/n Split dispatches between multiple command buffers. Track previous semaphore in context. (#12523)
Stack from [ghstack](https://github.com/ezyang/ghstack) (oldest at
bottom):
* #12525
* __->__ #12523
* #12522
* #12519

This diff is the third part of a series of diffs aiming to split
dispatches between multiple command buffers. In this diff, we are
tracking the previous semaphore in the context.

A new member variable `prev_semaphore_` was added to the `Context`
class. This variable is used to store the semaphore of the previously
submitted command buffer.

Differential Revision:
[D78360037](https://our.internmc.facebook.com/intern/diff/D78360037/)

---------

Co-authored-by: Stephen Jia <ssjia@meta.com>