Make ExecuTorch build with Xcode 26 beta toolchain and InferSendableFromCaptures enabled (#12539)
Summary:
Fix a bunch of staging builds that stopped working after enabling
`InferSendableFromCaptures` globally. This is likely caused by a bug in
Xcode's 26 toolchain, although I've not been able to reproduce it in an
isolated sample.

Switching from `Array.init` to `{ Array($0) }` solves the issue and
keeps the code building without warnings in all toolchain versions.

Differential Revision: D78406265

Co-authored-by: Xavier Jurado Cristobal <xavierjurado@meta.com>
Co-authored-by: Anthony Shoumikhin <anthony@shoumikh.in>