Add top-level CMake backends target (#12689)
Add a top level backends CMake target, which includes all configured
backends. I validate this change by building a simple runner using the
executorch_backends target and verifying that it was able to build and
run an XNNPACK-delegated binary. Once the changes land in ExecuTorch,
this will be long-term validated in executorch-examples CI.

To allow the executor runner target to use the executorch_backends
target, I've also moved the Vulkan defs above the runner. This causes
issues with the vulkan_executor_runner target, so I'm updating the main
runner to include Vulkan and remove the Vulkan-specific one (subject to
@SS-JIA's signoff).

This is done in the context of top-level CMake targets, tracked in
https://github.com/pytorch/executorch/issues/12293.