Add top-level CMake extensions target (#12696)
Add a top level extensions CMake target, which includes all configured
extensions. I validate this change by building a simple runner using the
executorch_extensions target and verifying that it was able to build and
run using the module and tensor extension. Once the changes land in
ExecuTorch, this will be long-term validated in executorch-examples CI.

This is done in the context of top-level CMake targets, tracked in
https://github.com/pytorch/executorch/issues/12293.