Add top-level CMake kernels target (#12806)
Add a top level kernels CMake target, which includes all configured
kernel libraries. I validate this change by building a simple runner
using the executorch_kernels target and verifying that it was able to
build and run MobileNet v2, with and without selective build. Once the
changes land in ExecuTorch, this will be long-term validated in
executorch-examples CI.

There are a few follow-ups needed, which I will handle shortly. I still
need to add merging logic for the YAML files to generate a single
selective library that includes quantized and LLM ops, when configured
to build them.

This is done in the context of top-level CMake targets, tracked in
https://github.com/pytorch/executorch/issues/12293.