CI Tests for Zephyr Integration (#12734)
### Summary
Adding in CI tests to verify and maintain support for Zephyr projects
that build executorch as a module. The main zephyr application source
code is hosted externally for now, but the process to host an Executorch
as an official module of zephyr is in progress.

This PR runs the simple Add model as unit tests, and verifies that the
output is `2.0` (1D tensor + 1D tensor, values filled with `1.0s`). This
is intended to be a simple and fast test to exemplify how to use
executorch to build an executor_runner for other models.

### Test plan
Added a new CI test called `test-models-arm-zephyr` that exits with
status 0 on success, and 1 on failure.