Arm backend: Eliminate one memory copy in executor runner (#12992)
Prior to this patch, temporary tensors/buffers were used to hold the
input data which was then copied over to the actual input tensors for
running the inference. This patch removes this copying by instead
writing the input data directly to the input tensors.

Signed-off-by: per.held@arm.com


cc @digantdesai @freddan80 @per @zingo @oscarandersson8218

Signed-off-by: per.held@arm.com
Co-authored-by: Martin Lindström <Martin.Lindstroem@arm.com>