Change default minimum_deployment_target to None and support CoreML models with no inputs (#13053)
This PR changes the default minimum_deployment_target in CoreML from the
fixed iOS15 to None. This will automatically select the minimum
deployment target required to export the model.

A warning occurs after export telling the user the deploymnet target
selected, and directing them to specify a specific target if that is
what they need.

This is more in line with how CoreML standalone works.

In addition, this PR allows running CoreML models with no user inputs,
which requires CoreML deployment target iOS18 or higher.

This addresses:
* https://github.com/pytorch/executorch/issues/11719
* https://github.com/pytorch/executorch/issues/12906