[ET-VK] Fix Build Errors in Vulkan Backend (#13170)
This change fixes build issues that arose from the addition of 2D
reduction to the Vulkan backend:

https://github.com/pytorch/executorch/commit/815cf1b2ea029b524d5d74b57e154fd4cd559faa

```get_tensor()``` was moved to be a protected member of ComputeGraph a few minutes before the above commit got merged.

This change also slightly modifies op_registry.py to have a more conservative approach of allowing 2D reduction to be delegated.

cc @SS-JIA @manuelcandales @cbilgin