[ET-VK] 8/n Split dispatches between multiple command buffers. This diff adds a config to limit the maximum number of command buffers created when splitting execution between multiple command buffers. (#13204)
This diff adds a config to limit the maximum number of command buffers created when splitting execution between multiple command buffers."

This diff introduces a new configuration option, `execute_max_cmds`, to limit the maximum number of command buffers created when splitting execution between multiple command buffers. This feature allows for more efficient management of command buffers, particularly in scenarios where the number of nodes in the graph is large.

Differential Revision: [D79575908](https://our.internmc.facebook.com/intern/diff/D79575908/)