Always run build-presets workflow on PRs/pushes (#13243)
The build-presets workflow currently runs only when the workflow file
itself is updated. As we update docs to recommend using presets, we
should run this job on all PRs.

Note that the Windows preset is currently reporting success despite
failing to actually build. I'm intending to resolve that shortly as a
follow-up.