Update build from source docs (#13210)
### Summary
Overhaul the "Building from Source" doc page. The primarily intent of
these changes is to document CMake presets and the various build options
that we expose. However, I also did a pass on the existing contents of
the file to improve formatting and clarity.

I've re-organized the page to clearly delineate environment setup,
python install, and native build. It should flow better and be easier to
read.

### Test plan
I have built the docs locally to inspect the contents for formatting and
correctness.

Preview page:
https://docs-preview.pytorch.org/pytorch/executorch/13210/using-executorch-building-from-source.html
Live page (for comparison):
https://docs.pytorch.org/executorch/0.7/using-executorch-building-from-source.html