Install NXP and OpenVINO backends, top-level CMake targets (#13673)
Make the top-level backends, extensions, and kernels targets
installable. See https://github.com/pytorch/executorch/issues/12949. Add
CMake export directives for NXP and OpenVINO, which are needed to
support the exportable top-level targets.

I've also removed the apple and llm/apple extensions from the
executorch_extensions target. From what I understand, these won't be
consumed this way, as they are build via apple frameworks. There is some
installation complexity to make it work with install, so I'm inclined to
skip it if it won't be used. Will check with Anthony in the comments.