Exynos Backend for Executorch to bring up on Exynos SoC (#13677)
### Summary
This PR introduces support for the Exynos backend of Executorch, can be
inferenced on Exynos acceletors.
Exynos Backend is a crucial component designed to optimize the
performance of AI models on Exynos chipsets. Leveraging Exynos AI
LiteCore, it enables comprehensive support for NPU/DSP, CPU, and GPU,
ensuring seamless execution of AI workloads on Exynos-based devices.

Below is a description of the features:
* Exynos Backend : The backends/samsung directory were made by ENN
Partitioner, I/F for communication with Exynos LiteCore, runtime backend
files and device-specific optimization. See backend/samsung/README.md
for usage

* Exynos examples: The examples/samsung directory privides scripts for
AOT compilation, quantization and so on. please refer
examples/samsung/README.md


### Test plan
This PR is tested with Exynos Backend on NPU/DSP/CPU/GPU of Exynos 2500
SoC. To run the layer and models tests with README.md


cc: @mergennachin @SS-JIA  @digantdesai @kimishpatel

---------

Signed-off-by: jiseong.oh <jiseong.oh@samsung.com>
Signed-off-by: chong-chen <chong.chen@samsung.com>
Co-authored-by: chong-chen <chong.chen@samsung.com>
Co-authored-by: Hoon-Choi <hoon98.choi@samsung.com>