Add EXECUTORCH_THREADPOOL_SIZE options, default to using only performance cores (#14090)
### Summary
Allow build-time configuration of the thread pool size and default to a
performance heuristic.

There are 2 modes that we want to support:
* Heuristic-based. Choose the number of threads according to a
performance heuristic. Use threads equal to the number of detected
performance cores, but we can continue to iterate on this by adding
fine-grained heuristics for specific chipsets in the future.
* All cores (threads=cores). This is the current behavior. We need to
maintain this as an option for some use cases.

With this PR, the default (for OSS) is to use performance cores. From
testing with CV models on ~10 representative devices across the
performance spectrum, this gives anywhere from parity with the existing
perf to up to a 13x speedup (measured on Pixel 6). Many common devices
(S20, S22, iPhone 15 Pro) show a 2-4x speedup.

#### Specifying Threadpool Size

To specify the threadpool size, I've added two preprocessor options (and
corresponding CMake options):
* `EXECUTORCH_THREADPOOL_USE_PERFORMANCE_CORES`- Use threads = detected
perf cores.
* `EXECUTORCH_THREADPOOL_USE_ALL_LOGICAL_CORES` - Use threads = logical
cores.

### Test plan
I've verified that logic functions correctly in OSS by building the
executor_runner on M1 Mac and observing the existing logging in
cpuinfo_utils.

Measuring MobileNet V3 (exported from examples) on XNNPACK, time to run
100 iterations drops from ~450ms to ~230ms on M1 Pro with this change.