[Backend Tester] Add flow-specific skipped tests, skip argmin/max on Core ML due to native crashes, float16/64 on Vulkan (#13992)
Argmin and argmax are crashing on Core ML. Native crashes are hard to
recover from. In the medium term, I'd like to look at using
pytest-xdist, which, by nature of process isolation, can recover. But
for now, I'm adding a way to skip tests on specific backend flows to
allow running the remaining tests. I've also skipped Vulkan float16 and
float64 tests, as they crash when using swiftshader (used in the CI for
software execution of the Vulkan compute shaders). This should be
re-visited in the future.

There is also a little technical debt in the code duplication between
operator and model tests. This will also be cleaned up shortly.