Cortex_m backend: Add cortex_m tester + test_add (#14510)
Note that tests are currently failing, comparing for example the call to
arm_elementwise_add_s8 in op_quantized_add.cpp
https://github.com/pytorch/executorch/blob/e7713d3210349b7acdc8f9e79837842efea34dcf/backends/cortex_m/ops/op_quantized_add.cpp#L88
with the definition in CMSIS-NN

https://github.com/ARM-software/CMSIS-NN/blob/88f1982a69c00ed13dd633a63da1009c48abbb4d/Include/arm_nnfunctions.h#L1923
it seems that the args are listed in the wrong order. This will be fixed
in a future patch.

Minor fixes to get this to work:
- Add init file to make test names unique
- Update conftest to not crash is_option_enabled for tests running from
external folder

Signed-off-by: Adrian Lundell <adrian.lundell@arm.com>